+++ jvm/src/main/scala/gpdviz/async/NullNotifier.scala
@@ -7,7 +7,8 @@ object NullNotifier extends Notifier {
 
   def details: String = "no-notifier"
 
-  def getSensorSystemIndex(sysid: String, ssOpt: Option[SensorSystem]): String = ""
+  def getSensorSystemIndex(sysid: String, ssOpt: Option[SensorSystem],
+                           indexResource: String = "web/index.html"): String = ""
 
   def notifySensorSystemRegistered(ss: SensorSystem): Unit = ()